split integration tests for model cleanup
